Update Dockerfile to use working release of VTK
The release versions of VTK have caught up to the development branch that introduces definitions regarding vtknlohmann. These definitions are causing conflicts with my own linking to nlohmannjson. Will need to spend some time looking at how to resolve these with newer versions of VTK. 

As a temporary (potentially permanent) measure I have opted to checkout an older version of VTK (v9.3.1) as I know it works and I do not make use of any of the newer VTK features.